use a single DashMetricsMock definition
