+++ test/unit/mocks/DashMetricsMock.js
@@ -1,0 +1,14 @@
+class DashMetricsMock {
+    constructor() {
+    }
+
+    getCurrentDVRInfo() {
+        return null;
+    }
+
+    getCurrentBufferLevel() {
+        return 15;
+    }
+}
+
+export default DashMetricsMock;
